refactor: implementacion de Gradle y JavaFX
