+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Wed Oct 23 23:18:35 BOT 2024
+gradle.version=8.8
